+++ uefi/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 - Exported `data_types::FromSliceUntilNulError`.
 - Added `proto::console::pointer::AbsolutePointer` protocol.
+- Added `CString::clear` and `PathBuf::clear` functions.
 
 ## Changed
 - **Breaking**: Changed `Server::server_type` and the `server_type` parameter